renew and submit book added

+++ src/controllers/issue_controller.py
@@ -1,5 +1,7 @@
 from src.db_manager import db_manager
 from datetime import datetime
+from src.controllers.book_controller import BookController
+from src.controllers.member_controller import MemberController
 
 class IssueController:
     def __init__(self):
@@ -7,6 +9,20 @@ class IssueController:
 
     def issue_book(self, book_id, member_id):
         """Issues a book to a member."""
+
+        book_status = BookController().get_book_by_id(book_id)
+    
+        if not book_status:
+            return {"success": False, "message": "Book not found in the system!"}
+
+        if not book_status["isAvail"]: 
+            return {"success": False, "message": "Book is already issued and not available!"}
+
+        member_exists = MemberController().get_member_by_id(member_id)
+        
+        if not member_exists:
+            return {"success": False, "message": "Member not found! Cannot issue book."}
+
         query_issue = "INSERT INTO tbl_issue (bookID, memberID, issueTime, renewCount) VALUES (?, ?, ?, ?)"
         query_update = "UPDATE tbl_addbook SET isAvail = FALSE WHERE id = ?"
 
@@ -20,6 +36,15 @@ class IssueController:
 
     def return_book(self, book_id):
         """Returns a book and updates availability."""
+
+        book_status = BookController().get_book_by_id(book_id)
+    
+        if not book_status:
+            return {"success": False, "message": "Book not found in the system!"}
+
+        if book_status["isAvail"]: 
+            return {"success": False, "message": "Book needs to be issued to be returned"}
+        
         query_delete = "DELETE FROM tbl_issue WHERE bookID = ?"
         query_update = "UPDATE tbl_addbook SET isAvail = TRUE WHERE id = ?"
 
@@ -29,6 +54,15 @@ class IssueController:
 
     def renew_book(self, book_id):
         """Renews the issued book."""
+
+        book_status = BookController().get_book_by_id(book_id)
+    
+        if not book_status:
+            return {"success": False, "message": "Book not found in the system!"}
+
+        if book_status["isAvail"]: 
+            return {"success": False, "message": "Book is not issued to be renewed."}
+        
         query = "UPDATE tbl_issue SET issueTime = ?, renewCount = renewCount + 1 WHERE bookID = ?"
         values = (datetime.now(), book_id)
         self.db.execute_query(query, values)
@@ -37,4 +71,7 @@ class IssueController:
     def get_issued_books(self):
         """Fetches all issued books."""
         query = "SELECT * FROM tbl_issue"
-        return self.db.fetch_data(query)
+        books = self.db.fetch_data(query)
+        if books is not None and len(books) == 0:
+            return  {"success": False, "message": "No books issued"}
+        return  {"success": True, "books": books, "message": "All books fetched!"}
